Move styles to a seperate component
